[FLINK-10247][metric]Pass the address of Metric Query Service actor system to TaskExecutor and fix up
In order to making JobManager's metric query service communicate with Task executor's, JobManager should be
aware of the address of TaskExecutor's MetricQueryService actor. This commit changes the constructor
of TaskExecutor by passing the path of MetricQueryService actor and save it in an instance variable. Meanwhile,
it changes the TaskExecutorGateway interface to make the address available to JobManager.

Besides changes mentioned above, this commit also add an option to configure the port range of metric query
service in JobManager started in a Yarn cluster.