[FLINK-10687] [table] Introduce flink-table-common
Because more and more table factories for connectors and formats are
added and external catalog support is also on the horizon, flink-table
becomes a dependency for many Flink modules. Since flink-table is
implemented in Scala it requires other modules to be suffixes with
Scala prefixes. However, as we have learned in the past, Scala code
is hard to maintain which is why our long-term goal is to avoid
Scala/Scala dependencies.

This commits adds a new module flink-table-common that contains
interfaces between flink-table and other modules. This module is
implemented in Java and should contain minimal (or better no)
external dependencies.

(cherry picked from commit f3bce3c)