[FLINK-10848] Remove container requests after successful container allocation
This commit removes container requests after containers have been allocated. This prevents that
we will request more and more containers from Yarn in case of a recovery.

Since we cannot rely on the reported container Resource, we remove the container request by
using the requested Resource. This is due Yarn's DefaultResourceCalculator which neglects the
number of vCores when allocating containers.

This closes #7430.

(cherry picked from commit 3903b81)